requirements.txt
spidev>=3.5
Pillow>=8.0.0
requests>=2.25.0
